[BREAKING] Raise exceptions when accessing property in not-loaded family
Previously, properties in not-loaded families were returned with default values.

This brought all sorts of issues with over-writing data in not-loaded families
with these default values, unnecessary complexity, etc.

Now, when you try to access a property in a not-loaded family, an FamilyNotLoaded
exception is raised, to prevent you from losing data:

    article = Article.find(1)
    article.views
    # => FamilyNotLoaded

Notice that you can freely *set* values for not loaded properties, as in:

    article = Article.find(1)
    article.views = 100